\docType{data}

\name{prostate}

\alias{prostate}

\title{Clinical Data of a Cohort of Prostate Cancer Patiens}

\description{The data belong to a cohort of 35 patients with prostate cancer from two different hospitals.
}

\usage{data(prostate)}

\value{
The data.frame "\code{prostate}" with the following elements: "\code{Hospital}", "\code{Gender}", "\code{Gleason score}",  "\code{BMI}", and "\code{Age}".
}

\examples{
data(prostate)

head(prostate)
}

\keyword{datasets}
